backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-environ>=0.11
django-cors-headers>=4.3


pandas>=2.0
numpy>=2.0

celery>=5.3
redis>=5.0

yfinance

pyarrow>=15.0
